+++ libs/forms/src/lib/utils/index.ts
@@ -1,2 +1,3 @@
 export { deepEqual } from './deep-equal'
 export { safeStringify } from './safe-stringify'
+export { useFormStoreSubscribe } from './use-form-store-subscribe'